# Conflicts:
#   src/main/java/com/nsa/clinical/controllers/QuestionController.java
#   src/main/java/com/nsa/clinical/controllers/QuestionnaireController.java
#   src/main/java/com/nsa/clinical/entities/Question.java

+++ src/main/java/com/nsa/clinical/repositories/QuestionRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * Created by c1673142 on 18-Nov-17.
  */
 public interface QuestionRepository extends JpaRepository<Question,Long> {
-    Question findByQuestionId(int id);
+    Question findByQuestionId(long id);
 }
